Patch [1424156]: Fixes compiler warnings
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@1964 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/cbproject.cpp
@@ -124,7 +124,7 @@ void cbProject::NotifyPlugins(wxEventType type)
 
 void cbProject::SetCompilerIndex(int compilerIdx)
 {
-    if(abs(compilerIdx)>=CompilerFactory::Compilers.GetCount())
+    if(((unsigned int) abs(compilerIdx)) >= CompilerFactory::Compilers.GetCount())
         return; // Invalid compiler
     if (compilerIdx != m_CompilerIdx)
     {
